Add new chord app and move chord-specific stuff to it
