+++ chord_metadata_service/chord/admin.py
@@ -1,0 +1,18 @@
+from django.contrib import admin
+
+from .models import *
+
+
+@admin.register(Project)
+class ProjectAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(Dataset)
+class DatasetAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(TableOwnership)
+class TableOwnershipAdmin(admin.ModelAdmin):
+    pass
