+++ chord_metadata_service/chord/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ChordConfig(AppConfig):
+    name = 'chord'
