+++ chord_metadata_service/metadata/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 from chord_metadata_service.restapi import api_views, urls as restapi_urls
-from chord_metadata_service.patients import chord_ingest_views, chord_search_views
+from chord_metadata_service.chord import views_ingest, views_search
 from rest_framework.schemas import get_schema_view
 
 from .settings import DEBUG
@@ -34,14 +34,14 @@ urlpatterns = [
     path('api/', include(restapi_urls)),
     path('service-info/', api_views.service_info),
 
-    path('workflows', chord_ingest_views.workflow_list),
-    path('workflows/<slug:workflow_id>', chord_ingest_views.workflow_item),
-    path('workflows/<slug:workflow_id>.wdl', chord_ingest_views.workflow_file),
+    path('workflows', views_ingest.workflow_list),
+    path('workflows/<slug:workflow_id>', views_ingest.workflow_item),
+    path('workflows/<slug:workflow_id>.wdl', views_ingest.workflow_file),
 
-    path('ingest', chord_ingest_views.ingest),
+    path('ingest', views_ingest.ingest),
 
-    path('data-types', chord_search_views.data_type_list),
-    path('datasets', chord_search_views.dataset_list),
-    path('datasets/<str:dataset_id>', chord_search_views.dataset_detail),
-    path('search', chord_search_views.chord_search),
+    path('data-types', views_search.data_type_list),
+    path('datasets', views_search.dataset_list),
+    path('datasets/<str:dataset_id>', views_search.dataset_detail),
+    path('search', views_search.chord_search),
 ] + [path('admin/', admin.site.urls)] if DEBUG else []
