+++ chord_metadata_service/phenopackets/serializers.py
@@ -239,46 +239,3 @@ class InterpretationSerializer(GenericSerializer):
 	class Meta:
 		model = Interpretation
 		fields = '__all__'
-
-
-#############################################################
-#                                                           #
-#              Project Management  Serializers              #
-#                                                           #
-#############################################################
-
-class ProjectSerializer(GenericSerializer):
-	# noinspection PyMethodMayBeStatic
-	def validate_data_use(self, value):
-		validation = Draft7Validator(CHORD_DATA_USE_SCHEMA).is_valid(value)
-		if not validation:
-			raise serializers.ValidationError("Data use is not valid")
-		return value
-
-	# noinspection PyMethodMayBeStatic
-	def validate_name(self, value):
-		if len(value.strip()) < 3:
-			raise serializers.ValidationError("Name must be at least 3 characters")
-		return value.strip()
-
-	class Meta:
-		model = Project
-		fields = '__all__'
-
-
-class DatasetSerializer(GenericSerializer):
-	# noinspection PyMethodMayBeStatic
-	def validate_name(self, value):
-		if len(value.strip()) < 3:
-			raise serializers.ValidationError("Name must be at least 3 characters")
-		return value.strip()
-
-	class Meta:
-		model = Dataset
-		fields = '__all__'
-
-
-class TableOwnershipSerializer(GenericSerializer):
-	class Meta:
-		model = TableOwnership
-		fields = '__all__'
