+++ chord_metadata_service/restapi/urls.py
@@ -3,6 +3,7 @@ from django.urls import path, include
 from rest_framework import routers
 # from chord_metadata_service.patients import api_views, chord_api_views
 # from rest_framework.schemas import get_schema_view
+from chord_metadata_service.chord import api_views as chord_views
 from chord_metadata_service.patients import api_views as individual_views
 from chord_metadata_service.phenopackets import api_views as phenopacket_views
 
@@ -30,9 +31,9 @@ router.register(r'phenopackets', phenopacket_views.PhenopacketViewSet)
 router.register(r'genomicinterpretations', phenopacket_views.GenomicInterpretationViewSet)
 router.register(r'diagnoses', phenopacket_views.DiagnosisViewSet)
 router.register(r'interpretations', phenopacket_views.InterpretationViewSet)
-router.register(r'projects', phenopacket_views.ProjectViewSet)
-router.register(r'datasets', phenopacket_views.DatasetViewSet)
-router.register(r'table_ownership', phenopacket_views.TableOwnershipViewSet)
+router.register(r'projects', chord_views.ProjectViewSet)
+router.register(r'datasets', chord_views.DatasetViewSet)
+router.register(r'table_ownership', chord_views.TableOwnershipViewSet)
 
 urlpatterns = [
 	path('', include(router.urls)),
